+++ lib/simple_3d.dart
@@ -1,4 +1,4 @@
-export 'src/enum_sp3d_drawing_mode.dart';
+export 'src/enum_sp3d_draw_mode.dart';
 export 'src/sp3d_face.dart';
 export 'src/sp3d_fragment.dart';
 export 'src/sp3d_material.dart';