+++ lib/src/sp3d_obj.dart
@@ -30,7 +30,7 @@ class Sp3dObj {
   Sp3dPhysics? physics;
   Map<String, dynamic>? option;
   int layerNum;
-  EnumSp3dDrawingMode drawingMode;
+  EnumSp3dDrawMode drawMode;
 
   /// Constructor
   /// * [vertices] : vertex list.
@@ -44,7 +44,7 @@ class Sp3dObj {
   /// * [option] : Optional attributes that may be added for each app.
   /// * [layerNum] : If the renderer has layerNum enabled,
   /// it forces the drawing order to be in ascending order of this value.
-  /// * [drawingMode] : Specifies the drawing mode when rendering.
+  /// * [drawMode] : Specifies the drawing mode when rendering.
   /// Specify rect when speed is required but 3D structure is not required,
   /// such as scatter plots.
   Sp3dObj(this.vertices, this.fragments, this.materials, this.images,
@@ -54,7 +54,7 @@ class Sp3dObj {
       this.physics,
       this.option,
       this.layerNum = 0,
-      this.drawingMode = EnumSp3dDrawingMode.normal});
+      this.drawMode = EnumSp3dDrawMode.normal});
 
   /// Deep copy the object.
   Sp3dObj deepCopy() {
@@ -81,7 +81,7 @@ class Sp3dObj {
         physics: physics != null ? physics!.deepCopy() : null,
         option: option != null ? {...option!} : null,
         layerNum: layerNum,
-        drawingMode: drawingMode);
+        drawMode: drawMode);
   }
 
   /// Convert the object to a dictionary.
@@ -123,7 +123,7 @@ class Sp3dObj {
     d['physics'] = physics != null ? physics!.toDict() : null;
     d['option'] = option;
     d['layer_num'] = layerNum;
-    d['drawing_mode'] = drawingMode.index;
+    d['draw_mode'] = drawMode.index;
     return d;
   }
 
@@ -167,9 +167,9 @@ class Sp3dObj {
             : null,
         option: src['option'],
         layerNum: src.containsKey('layer_num') ? src['layer_num'] : 0,
-        drawingMode: src.containsKey('drawing_mode')
-            ? EnumSp3dDrawingMode.values[src['drawing_mode']]
-            : EnumSp3dDrawingMode.normal);
+        drawMode: src.containsKey('draw_mode')
+            ? EnumSp3dDrawMode.values[src['draw_mode']]
+            : EnumSp3dDrawMode.normal);
   }
 
   /// (en)Adds the specified vector to all vectors of this object.
@@ -198,11 +198,11 @@ class Sp3dObj {
   }
 
   /// (en)Merge another object into this object. This operation is high cost.
-  /// id, name, author, physics, option, layerNum and drawingMode values do not change.
+  /// id, name, author, physics, option, layerNum and drawMode values do not change.
   ///
   /// (ja)このオブジェクトに別のオブジェクトをマージします。この操作は高コストです。
   /// このオブジェクト固有のパラメータ（id,name,author,physics）とオプション値、
-  /// layerNum、drawingModeは変更されません。
+  /// layerNum、drawModeは変更されません。
   ///
   /// * [other] : other obj.
   Sp3dObj merge(Sp3dObj other) {